hardware/alsa_sound: Miscellaneous fixes in device routing
Switch device function need not be called for every usecase, ucm
takes care of switching the devices for all use cases if device
is switched for any one of the use cases. Avoid calling this
function for every use case in doRouting function.

With current logic device value to be disabled during session
closure may not be correct sometimes if previous session uses
tty or dualmic devices. This is because the device value is
computed based on the setParameter values of tty/dualmic which
might have been changed by the time device is disabled and
can result in disabling of wrong devices. Instead change the
logic to use cached Rx and Tx ucm device values.

Voice calls require switching of both Rx and Tx mixer controls
on device switch though only Rx or Tx device is switched. If
mode is set to in-call mode call switch device ucm API for both
Rx and Tx even if only one of them is changed.

CRs-fixed: 298483
Change-Id: Ifeeae3ec50df3fa39449d5683650ec41fbec1070