hardware/qcom/alsa_sound: run-time calculate number of periods
Get the number of periods by calculating it based on
the buffer size & period size, provided by the driver.
Since the number of periods is already hard-coded inside the driver,
it doesn't make sense that AHAL assigns it, again and cause the set
hardware param failure because of mismatching of the parameter.

Change-Id: Ic271f4e705c4b01a82f79c39aff9769cbb29892a