hardware/alsa_sound: Update the method to check the Tabla rev.
The current implementation to use the cpu info to determine the Tabla
Codec revision has the limitation because it assumes the newer version
of  msm is tied to the newer version of the Tabla. The fix is to use
/proc/asound/cards to determine the Tabla revision regardless the cpu
info.

CRs-fixed: 329058
(cherry picked from commit 13832e8a4b7da9e13212b9590694caa15f848d0b)

Change-Id: If88083762711cbf8ff834869652eabffce3c9e82